@@ -33,9 +33,7 @@ from aesara.tensor.nlinalg import det, eigh, matrix_inverse, trace
 from aesara.tensor.random.basic import dirichlet, multinomial, multivariate_normal
 from aesara.tensor.random.op import RandomVariable, default_supp_shape_from_params
 from aesara.tensor.random.utils import broadcast_params, normalize_size_param
-from aesara.tensor.slinalg import Cholesky
-from aesara.tensor.slinalg import solve_lower_triangular as solve_lower
-from aesara.tensor.slinalg import solve_upper_triangular as solve_upper
+from aesara.tensor.slinalg import Cholesky, SolveTriangular
 from aesara.tensor.type import TensorType
 from scipy import linalg, stats
 
@@ -43,12 +41,7 @@ import pymc as pm
 
 from pymc.aesaraf import change_rv_size, floatX, intX
 from pymc.distributions import transforms
-from pymc.distributions.continuous import (
-    BoundedContinuous,
-    ChiSquared,
-    Normal,
-    assert_negative_support,
-)
+from pymc.distributions.continuous import BoundedContinuous, ChiSquared, Normal
 from pymc.distributions.dist_math import (
     betaln,
     check_parameters,
@@ -83,6 +76,9 @@ __all__ = [
     "CAR",
     "StickBreakingWeights",
 ]
+
+solve_lower = SolveTriangular(lower=True)
+solve_upper = SolveTriangular(lower=False)
 
 
 class SimplexContinuous(Continuous):
@@ -356,42 +352,49 @@ class MvStudentT(Continuous):
     nu : tensor_like of float
         Degrees of freedom, should be a positive scalar.
     Sigma : tensor_like of float, optional
-        Covariance matrix. Use `cov` in new code.
+        Scale matrix. Use `scale` in new code.
     mu : tensor_like of float, optional
         Vector of means.
-    cov : tensor_like of float, optional
-        The covariance matrix.
+    scale : tensor_like of float, optional
+        The scale matrix.
     tau : tensor_like of float, optional
         The precision matrix.
     chol : tensor_like of float, optional
-        The cholesky factor of the covariance matrix.
+        The cholesky factor of the scale matrix.
     lower : bool, default=True
         Whether the cholesky fatcor is given as a lower triangular matrix.
     """
     rv_op = mv_studentt
 
     @classmethod
-    def dist(cls, nu, Sigma=None, mu=None, cov=None, tau=None, chol=None, lower=True, **kwargs):
+    def dist(cls, nu, Sigma=None, mu=None, scale=None, tau=None, chol=None, lower=True, **kwargs):
+        if kwargs.get("cov") is not None:
+            warnings.warn(
+                "Use the scale argument to specify the scale matrix."
+                "cov will be removed in future versions.",
+                FutureWarning,
+            )
+            scale = kwargs.pop("cov")
         if Sigma is not None:
-            if cov is not None:
-                raise ValueError("Specify only one of cov and Sigma")
-            cov = Sigma
+            if scale is not None:
+                raise ValueError("Specify only one of scale and Sigma")
+            scale = Sigma
         nu = at.as_tensor_variable(floatX(nu))
         mu = at.as_tensor_variable(floatX(mu))
-        cov = quaddist_matrix(cov, chol, tau, lower)
+        scale = quaddist_matrix(scale, chol, tau, lower)
         # Aesara is stricter about the shape of mu, than PyMC used to be
-        mu = at.broadcast_arrays(mu, cov[..., -1])[0]
-        assert_negative_support(nu, "nu", "MvStudentT")
-        return super().dist([nu, mu, cov], **kwargs)
+        mu = at.broadcast_arrays(mu, scale[..., -1])[0]
 
-    def moment(rv, size, nu, mu, cov):
+        return super().dist([nu, mu, scale], **kwargs)
+
+    def moment(rv, size, nu, mu, scale):
         moment = mu
         if not rv_size_is_none(size):
             moment_size = at.concatenate([size, [mu.shape[-1]]])
             moment = at.full(moment_size, moment)
         return moment
 
-    def logp(value, nu, mu, cov):
+    def logp(value, nu, mu, scale):
         """
         Calculate log-probability of Multivariate Student's T distribution
         at specified value.
@@ -405,7 +408,7 @@ class MvStudentT(Continuous):
         -------
         TensorVariable
         """
-        quaddist, logdet, ok = quaddist_parse(value, mu, cov)
+        quaddist, logdet, ok = quaddist_parse(value, mu, scale)
         k = floatX(value.shape[-1])
 
         norm = gammaln((nu + k) / 2.0) - gammaln(nu / 2.0) - 0.5 * k * at.log(nu * np.pi)
@@ -520,7 +523,7 @@ class Multinomial(Discrete):
     p : tensor_like of float
         Probability of each one of the different outcomes (0 <= p <= 1). The number of
         categories is given by the length of the last axis. Elements are expected to sum
-        to 1 along the last axis, and they will be automatically rescaled otherwise.
+        to 1 along the last axis.
     """
     rv_op = multinomial
 
@@ -544,7 +547,7 @@ class Multinomial(Discrete):
         n = at.shape_padright(n)
         mode = at.round(n * p)
         diff = n - at.sum(mode, axis=-1, keepdims=True)
-        inc_bool_arr = at.abs_(diff) > 0
+        inc_bool_arr = at.abs(diff) > 0
         mode = at.inc_subtensor(mode[inc_bool_arr.nonzero()], diff[inc_bool_arr.nonzero()])
         if not rv_size_is_none(size):
             output_size = at.concatenate([size, [p.shape[-1]]])
@@ -838,7 +841,7 @@ class PosDefMatrix(Op):
     def make_node(self, x):
         x = at.as_tensor_variable(x)
         assert x.ndim == 2
-        o = TensorType(dtype="int8", broadcastable=[])()
+        o = TensorType(dtype="int8", shape=[])()
         return Apply(self, [x], [o])
 
     # Python implementation:
@@ -880,7 +883,7 @@ class WishartRV(RandomVariable):
     @classmethod
     def rng_fn(cls, rng, nu, V, size):
         scipy_size = size if size else 1  # Default size for Scipy's wishart.rvs is 1
-        result = stats.wishart.rvs(np.int(nu), V, size=scipy_size, random_state=rng)
+        result = stats.wishart.rvs(int(nu), V, size=scipy_size, random_state=rng)
         if size == (1,):
             return result[np.newaxis, ...]
         else:
@@ -918,7 +921,7 @@ class Wishart(Continuous):
     ----------
     nu : tensor_like of int
         Degrees of freedom, > 0.
-    V : array_like
+    V : tensor_like of float
         p x p positive definite matrix.
 
     Notes
@@ -1262,27 +1265,27 @@ class LKJCholeskyCov:
 
     Parameters
     ----------
-    name: str
+    name : str
         The name given to the variable in the model.
-    eta: float
+    eta : tensor_like of float
         The shape parameter (eta > 0) of the LKJ distribution. eta = 1
         implies a uniform distribution of the correlation matrices;
         larger values put more weight on matrices with few correlations.
-    n: int
+    n : tensor_like of int
         Dimension of the covariance matrix (n > 1).
-    sd_dist: unnamed distribution
+    sd_dist : Distribution
         A positive scalar or vector distribution for the standard deviations, created
         with the `.dist()` API. Should have `shape[-1]=n`. Scalar distributions will be
         automatically resized to ensure this.
 
         .. warning:: sd_dist will be cloned, rendering it independent of the one passed as input.
 
-    compute_corr: bool, default=True
+    compute_corr : bool, default=True
         If `True`, returns three values: the Cholesky decomposition, the correlations
         and the standard deviations of the covariance matrix. Otherwise, only returns
         the packed Cholesky decomposition. Defaults to `True`.
         compatibility.
-    store_in_trace: bool, default=True
+    store_in_trace : bool, default=True
         Whether to store the correlations and standard deviations of the covariance
         matrix in the posterior trace. If `True`, they will automatically be named as
         `{name}_corr` and `{name}_stds` respectively. Effective only when
@@ -1290,13 +1293,13 @@ class LKJCholeskyCov:
 
     Returns
     -------
-    chol:  TensorVariable
+    chol :  TensorVariable
         If `compute_corr=True`. The unpacked Cholesky covariance decomposition.
-    corr: TensorVariable
+    corr : TensorVariable
         If `compute_corr=True`. The correlations of the covariance matrix.
-    stds: TensorVariable
+    stds : TensorVariable
         If `compute_corr=True`. The standard deviations of the covariance matrix.
-    packed_chol: TensorVariable
+    packed_chol : TensorVariable
         If `compute_corr=False` The packed Cholesky covariance decomposition.
 
     Notes
@@ -1507,9 +1510,9 @@ class LKJCorr(BoundedContinuous):
 
     Parameters
     ----------
-    n: int
+    n : tensor_like of int
         Dimension of the covariance matrix (n > 1).
-    eta: float
+    eta : tensor_like of float
         The shape parameter (eta > 0) of the LKJ distribution. eta = 1
         implies a uniform distribution of the correlation matrices;
         larger values put more weight on matrices with few correlations.
@@ -1649,20 +1652,20 @@ class MatrixNormal(Continuous):
 
     Parameters
     ----------
-    mu: array
+    mu : tensor_like of float
         Array of means. Must be broadcastable with the random variable X such
-        that the shape of mu + X is (m,n).
-    rowcov: mxm array
+        that the shape of mu + X is (M, N).
+    rowcov : (M, M) tensor_like of float, optional
         Among-row covariance matrix. Defines variance within
         columns. Exactly one of rowcov or rowchol is needed.
-    rowchol: mxm array
+    rowchol : (M, M) tensor_like of float, optional
         Cholesky decomposition of among-row covariance matrix. Exactly one of
         rowcov or rowchol is needed.
-    colcov: nxn array
+    colcov : (N, N) tensor_like of float, optional
         Among-column covariance matrix. If rowcov is the identity matrix,
         this functions as `cov` in MvNormal.
         Exactly one of colcov or colchol is needed.
-    colchol: nxn array
+    colchol : (N, N) tensor_like of float, optional
         Cholesky decomposition of among-column covariance matrix. Exactly one
         of colcov or colchol is needed.
 
@@ -1850,27 +1853,26 @@ class KroneckerNormal(Continuous):
     ========  ==========================
     Support   :math:`x \in \mathbb{R}^N`
     Mean      :math:`\mu`
-    Variance  :math:`K = \bigotimes K_i` + \sigma^2 I_N
+    Variance  :math:`K = \bigotimes K_i + \sigma^2 I_N`
     ========  ==========================
 
     Parameters
     ----------
-    mu: array
+    mu : tensor_like of float
         Vector of means, just as in `MvNormal`.
-    covs: list of arrays
+    covs : list of arrays
         The set of covariance matrices :math:`[K_1, K_2, ...]` to be
         Kroneckered in the order provided :math:`\bigotimes K_i`.
-    chols: list of arrays
+    chols : list of arrays
         The set of lower cholesky matrices :math:`[L_1, L_2, ...]` such that
         :math:`K_i = L_i L_i'`.
-    evds: list of tuples
+    evds : list of tuples
         The set of eigenvalue-vector, eigenvector-matrix pairs
         :math:`[(v_1, Q_1), (v_2, Q_2), ...]` such that
         :math:`K_i = Q_i \text{diag}(v_i) Q_i'`. For example::
 
             v_i, Q_i = at.nlinalg.eigh(K_i)
-
-    sigma: scalar, variable
+    sigma : scalar, optional
         Standard deviation of the Gaussian white noise.
 
     Examples
@@ -2102,19 +2104,20 @@ class CAR(Continuous):
 
     Parameters
     ----------
-    mu: array
+    mu : tensor_like of float
         Real-valued mean vector
-    W: Numpy matrix
+    W : (M, M) tensor_like of int
         Symmetric adjacency matrix of 1s and 0s indicating
-        adjacency between elements.
-    alpha: float or array
+        adjacency between elements. If possible, *W* is converted
+        to a sparse matrix, falling back to a dense variable.
+        :func:`~aesara.sparse.basic.as_sparse_or_tensor_variable` is
+        used for this sparse or tensorvariable conversion.
+    alpha : tensor_like of float
         Autoregression parameter taking values between -1 and 1. Values closer to 0 indicate weaker
         correlation and values closer to 1 indicate higher autocorrelation. For most use cases, the
-        support of alpha should be restricted to (0, 1)
-    tau: float or array
+        support of alpha should be restricted to (0, 1).
+    tau : tensor_like of float
         Positive precision variable controlling the scale of the underlying normal variates.
-    sparse: bool, default=False
-        Determines whether or not sparse computations are used
 
     References
     ----------
@@ -2267,9 +2270,9 @@ class StickBreakingWeights(SimplexContinuous):
 
     Parameters
     ----------
-    alpha: float
+    alpha : tensor_like of float
         Concentration parameter (alpha > 0).
-    K: int
+    K : tensor_like of int
         The number of "sticks" to break off from an initial one-unit stick. The length of the weight
         vector is K + 1, where the last weight is one minus the sum of all the first sticks.
 
@@ -2287,9 +2290,6 @@ class StickBreakingWeights(SimplexContinuous):
     def dist(cls, alpha, K, *args, **kwargs):
         alpha = at.as_tensor_variable(floatX(alpha))
         K = at.as_tensor_variable(intX(K))
-
-        assert_negative_support(alpha, "alpha", "StickBreakingWeights")
-        assert_negative_support(K, "K", "StickBreakingWeights")
 
         return super().dist([alpha, K], **kwargs)
 
