@@ -14,6 +14,7 @@
 import functools
 import itertools
 import sys
+import warnings
 
 import aesara
 import aesara.tensor as at
@@ -1699,12 +1700,14 @@ class TestMatchesScipy:
                 Bernoulli("x")
 
     def test_discrete_weibull(self):
-        check_logp(
-            DiscreteWeibull,
-            Nat,
-            {"q": Unit, "beta": NatSmall},
-            discrete_weibull_logpmf,
-        )
+        with warnings.catch_warnings():
+            warnings.filterwarnings("ignore", "divide by zero encountered in log", RuntimeWarning)
+            check_logp(
+                DiscreteWeibull,
+                Nat,
+                {"q": Unit, "beta": NatSmall},
+                discrete_weibull_logpmf,
+            )
         check_selfconsistency_discrete_logcdf(
             DiscreteWeibull,
             Nat,
@@ -1731,8 +1734,10 @@ class TestMatchesScipy:
         )
 
     def test_diracdeltadist(self):
-        check_logp(DiracDelta, I, {"c": I}, lambda value, c: np.log(c == value))
-        check_logcdf(DiracDelta, I, {"c": I}, lambda value, c: np.log(value >= c))
+        with warnings.catch_warnings():
+            warnings.filterwarnings("ignore", "divide by zero encountered in log", RuntimeWarning)
+            check_logp(DiracDelta, I, {"c": I}, lambda value, c: np.log(c == value))
+            check_logcdf(DiracDelta, I, {"c": I}, lambda value, c: np.log(value >= c))
 
     def test_zeroinflatedpoisson(self):
         def logp_fn(value, psi, mu):
@@ -2072,12 +2077,13 @@ class TestMatchesScipy:
 
     @pytest.mark.parametrize("n", [2, 3])
     def test_wishart(self, n):
-        check_logp(
-            Wishart,
-            PdMatrix(n),
-            {"nu": Domain([0, 3, 4, np.inf], "int64"), "V": PdMatrix(n)},
-            lambda value, nu, V: scipy.stats.wishart.logpdf(value, np.int(nu), V),
-        )
+        with pytest.warns(UserWarning, match="Wishart distribution can currently not be used"):
+            check_logp(
+                Wishart,
+                PdMatrix(n),
+                {"nu": Domain([0, 3, 4, np.inf], "int64"), "V": PdMatrix(n)},
+                lambda value, nu, V: scipy.stats.wishart.logpdf(value, int(nu), V),
+            )
 
     @pytest.mark.parametrize("x,eta,n,lp", LKJ_CASES)
     def test_lkjcorr(self, x, eta, n, lp):
@@ -2369,12 +2375,15 @@ class TestMatchesScipy:
 
     @pytest.mark.parametrize("n", [2, 3, 4])
     def test_orderedlogistic(self, n):
-        check_logp(
-            OrderedLogistic,
-            Domain(range(n), dtype="int64", edges=(None, None)),
-            {"eta": R, "cutpoints": Vector(R, n - 1)},
-            lambda value, eta, cutpoints: orderedlogistic_logpdf(value, eta, cutpoints),
-        )
+        with warnings.catch_warnings():
+            warnings.filterwarnings("ignore", "invalid value encountered in log", RuntimeWarning)
+            warnings.filterwarnings("ignore", "divide by zero encountered in log", RuntimeWarning)
+            check_logp(
+                OrderedLogistic,
+                Domain(range(n), dtype="int64", edges=(None, None)),
+                {"eta": R, "cutpoints": Vector(R, n - 1)},
+                lambda value, eta, cutpoints: orderedlogistic_logpdf(value, eta, cutpoints),
+            )
 
     @pytest.mark.parametrize("n", [2, 3, 4])
     def test_orderedprobit(self, n):
@@ -2621,6 +2630,7 @@ class TestMatchesScipy:
             {"mu": R, "sigma": Rplus, "steps": Nat},
             ref_logp,
             decimal=select_by_precision(float64=6, float32=1),
+            extra_args={"init_dist": Normal.dist(0, 100)},
         )
 
 
@@ -2630,8 +2640,14 @@ class TestBound:
     def test_continuous(self):
         with Model() as model:
             dist = Normal.dist(mu=0, sigma=1)
-            UnboundedNormal = Bound("unbound", dist, transform=None)
-            InfBoundedNormal = Bound("infbound", dist, lower=-np.inf, upper=np.inf, transform=None)
+            with warnings.catch_warnings():
+                warnings.filterwarnings(
+                    "ignore", "invalid value encountered in add", RuntimeWarning
+                )
+                UnboundedNormal = Bound("unbound", dist, transform=None)
+                InfBoundedNormal = Bound(
+                    "infbound", dist, lower=-np.inf, upper=np.inf, transform=None
+                )
             LowerNormal = Bound("lower", dist, lower=0, transform=None)
             UpperNormal = Bound("upper", dist, upper=0, transform=None)
             BoundedNormal = Bound("bounded", dist, lower=1, upper=10, transform=None)
@@ -2666,7 +2682,11 @@ class TestBound:
     def test_discrete(self):
         with Model() as model:
             dist = Poisson.dist(mu=4)
-            UnboundedPoisson = Bound("unbound", dist)
+            with warnings.catch_warnings():
+                warnings.filterwarnings(
+                    "ignore", "invalid value encountered in add", RuntimeWarning
+                )
+                UnboundedPoisson = Bound("unbound", dist)
             LowerPoisson = Bound("lower", dist, lower=1)
             UpperPoisson = Bound("upper", dist, upper=10)
             BoundedPoisson = Bound("bounded", dist, lower=1, upper=10)
@@ -2713,8 +2733,12 @@ class TestBound:
         msg = "Cannot transform discrete variable."
         with pm.Model() as m:
             x = pm.Poisson.dist(0.5)
-            with pytest.raises(ValueError, match=msg):
-                pm.Bound("bound", x, transform=pm.distributions.transforms.log)
+            with warnings.catch_warnings():
+                warnings.filterwarnings(
+                    "ignore", "invalid value encountered in add", RuntimeWarning
+                )
+                with pytest.raises(ValueError, match=msg):
+                    pm.Bound("bound", x, transform=pm.distributions.transforms.log)
 
         msg = "Given dims do not exist in model coordinates."
         with pm.Model() as m:
@@ -2783,8 +2807,12 @@ class TestBound:
     def test_array_bound(self):
         with Model() as model:
             dist = Normal.dist()
-            LowerPoisson = Bound("lower", dist, lower=[1, None], transform=None)
-            UpperPoisson = Bound("upper", dist, upper=[np.inf, 10], transform=None)
+            with warnings.catch_warnings():
+                warnings.filterwarnings(
+                    "ignore", "invalid value encountered in add", RuntimeWarning
+                )
+                LowerPoisson = Bound("lower", dist, lower=[1, None], transform=None)
+                UpperPoisson = Bound("upper", dist, upper=[np.inf, 10], transform=None)
             BoundedPoisson = Bound("bounded", dist, lower=[1, 2], upper=[9, 10], transform=None)
 
         first, second = joint_logp(LowerPoisson, [0, 0], sum=False)[0].eval()
@@ -3080,7 +3108,9 @@ def test_serialize_density_dist():
     with pm.Model():
         pm.Normal("x")
         y = pm.DensityDist("y", logp=func, random=random)
-        pm.sample(draws=5, tune=1, mp_ctx="spawn")
+        with warnings.catch_warnings():
+            warnings.filterwarnings("ignore", ".*number of samples.*", UserWarning)
+            pm.sample(draws=5, tune=1, mp_ctx="spawn")
 
     import cloudpickle
 
@@ -3116,7 +3146,7 @@ def test_distinct_rvs():
     [
         ("logp", r"pm.logp\(rv, x\)"),
         ("logcdf", r"pm.logcdf\(rv, x\)"),
-        ("random", r"rv.eval\(\)"),
+        ("random", r"pm.draw\(rv\)"),
     ],
 )
 def test_logp_gives_migration_instructions(method, newcode):
@@ -3288,9 +3318,9 @@ class TestLKJCholeskCov:
         with pm.Model() as m:
             sd_dist = pm.Exponential.dist(1, size=3)
             x = pm.LKJCholeskyCov("x", n=3, eta=1, sd_dist=sd_dist)
-        with pytest.warns(None) as record:
+        with warnings.catch_warnings():
+            warnings.simplefilter("error")
             m.logp()
-        assert not record
 
     @pytest.mark.parametrize(
         "sd_dist",
