@@ -24,7 +24,7 @@ from aeppl import factorized_joint_logprob
 from aeppl.abstract import assign_custom_measurable_outputs
 from aeppl.logprob import logcdf as logcdf_aeppl
 from aeppl.logprob import logprob as logp_aeppl
-from aeppl.transforms import TransformValuesOpt
+from aeppl.transforms import TransformValuesRewrite
 from aesara.graph.basic import graph_inputs, io_toposort
 from aesara.tensor.random.op import RandomVariable
 from aesara.tensor.subtensor import (
@@ -187,7 +187,7 @@ def joint_logp(
     # Else we assume we were given a single rv and respective value
     elif not isinstance(rv_values, Mapping):
         if len(var) == 1:
-            rv_values = {var[0]: at.as_tensor_variable(rv_values).astype(var[0].type)}
+            rv_values = {var[0]: at.as_tensor_variable(rv_values).astype(var[0].dtype)}
         else:
             raise ValueError("rv_values must be a dict if more than one var is requested")
 
@@ -231,7 +231,7 @@ def joint_logp(
                 if original_value_var is not None and hasattr(original_value_var.tag, "transform"):
                     transform_map[value_var] = original_value_var.tag.transform
 
-    transform_opt = TransformValuesOpt(transform_map)
+    transform_opt = TransformValuesRewrite(transform_map)
     temp_logp_var_dict = factorized_joint_logprob(
         tmp_rvs_to_values,
         extra_rewrites=transform_opt,
