@@ -565,7 +565,7 @@ def _logp_forw(point, out_vars, in_vars, shared):
         new_in_vars = []
         for in_var in in_vars:
             if in_var.dtype in discrete_types:
-                float_var = at.TensorType("floatX", in_var.broadcastable)(in_var.name)
+                float_var = at.TensorType("floatX", in_var.shape)(in_var.name)
                 new_in_vars.append(float_var)
                 replace_int_input[in_var] = at.round(float_var).astype(in_var.dtype)
             else:
