@@ -11,8 +11,7 @@
 #   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #   See the License for the specific language governing permissions and
 #   limitations under the License.
-
-import sys
+import warnings
 
 import aesara
 import aesara.tensor as at
@@ -27,9 +26,6 @@ import pymc as pm
 from pymc.ode import DifferentialEquation
 from pymc.ode.utils import augment_system
 from pymc.tests.helpers import fast_unstable_sampling_mode
-
-IS_FLOAT32 = aesara.config.floatX == "float32"
-IS_WINDOWS = sys.platform == "win32"
 
 
 def test_gradients():
@@ -368,7 +364,12 @@ class TestDiffEqModel:
             y = pm.LogNormal("y", mu=pm.math.log(forward), sigma=sigma, observed=yobs)
 
             with aesara.config.change_flags(mode=fast_unstable_sampling_mode):
-                idata = pm.sample(50, tune=0, chains=1)
+                with warnings.catch_warnings():
+                    warnings.filterwarnings("ignore", ".*number of samples.*", UserWarning)
+                    warnings.filterwarnings(
+                        "ignore", "invalid value encountered in log", RuntimeWarning
+                    )
+                    idata = pm.sample(50, tune=0, chains=1)
 
         assert idata.posterior["alpha"].shape == (1, 50)
         assert idata.posterior["y0"].shape == (1, 50)
@@ -399,7 +400,12 @@ class TestDiffEqModel:
             y = pm.LogNormal("y", mu=pm.math.log(forward), sigma=sigma, observed=yobs)
 
             with aesara.config.change_flags(mode=fast_unstable_sampling_mode):
-                idata = pm.sample(50, tune=0, chains=1)
+                with warnings.catch_warnings():
+                    warnings.filterwarnings("ignore", ".*number of samples.*", UserWarning)
+                    warnings.filterwarnings(
+                        "ignore", "invalid value encountered in log", RuntimeWarning
+                    )
+                    idata = pm.sample(50, tune=0, chains=1)
 
         assert idata.posterior["alpha"].shape == (1, 50)
         assert idata.posterior["beta"].shape == (1, 50)
@@ -441,7 +447,12 @@ class TestDiffEqModel:
             y = pm.LogNormal("y", mu=pm.math.log(forward), sigma=sigma, observed=yobs)
 
             with aesara.config.change_flags(mode=fast_unstable_sampling_mode):
-                idata = pm.sample(50, tune=0, chains=1)
+                with warnings.catch_warnings():
+                    warnings.filterwarnings("ignore", ".*number of samples.*", UserWarning)
+                    warnings.filterwarnings(
+                        "ignore", "invalid value encountered in log", RuntimeWarning
+                    )
+                    idata = pm.sample(50, tune=0, chains=1)
 
         assert idata.posterior["R"].shape == (1, 50)
         assert idata.posterior["sigma"].shape == (1, 50, 2)
@@ -482,7 +493,12 @@ class TestDiffEqModel:
             y = pm.LogNormal("y", mu=pm.math.log(forward), sigma=sigma, observed=yobs)
 
             with aesara.config.change_flags(mode=fast_unstable_sampling_mode):
-                idata = pm.sample(50, tune=0, chains=1)
+                with warnings.catch_warnings():
+                    warnings.filterwarnings("ignore", ".*number of samples.*", UserWarning)
+                    warnings.filterwarnings(
+                        "ignore", "invalid value encountered in log", RuntimeWarning
+                    )
+                    idata = pm.sample(50, tune=0, chains=1)
 
         assert idata.posterior["beta"].shape == (1, 50)
         assert idata.posterior["gamma"].shape == (1, 50)
