@@ -1,23 +1,33 @@
+import warnings
+
+from typing import Any, Callable, Dict, Optional
+from unittest import mock
+
 import aesara
 import aesara.tensor as at
+import arviz as az
 import jax
 import numpy as np
 import pytest
 
 from aesara.compile import SharedVariable
 from aesara.graph import graph_inputs
+from numpyro.infer import MCMC
 
 import pymc as pm
 
-from pymc.sampling_jax import (
-    _get_batched_jittered_initial_points,
-    _get_log_likelihood,
-    _replace_shared_variables,
-    get_jaxified_graph,
-    get_jaxified_logp,
-    sample_blackjax_nuts,
-    sample_numpyro_nuts,
-)
+with pytest.warns(UserWarning, match="module is experimental"):
+    from pymc.sampling_jax import (
+        _get_batched_jittered_initial_points,
+        _get_log_likelihood,
+        _numpyro_nuts_defaults,
+        _replace_shared_variables,
+        _update_numpyro_nuts_kwargs,
+        get_jaxified_graph,
+        get_jaxified_logp,
+        sample_blackjax_nuts,
+        sample_numpyro_nuts,
+    )
 
 
 @pytest.mark.parametrize(
@@ -106,9 +116,9 @@ def test_get_jaxified_graph():
     # be removed once https://github.com/aesara-devs/aesara/issues/637 is sorted.
     x = at.scalar("x")
     y = at.exp(x)
-    with pytest.warns(None) as record:
+    with warnings.catch_warnings():
+        warnings.simplefilter("error")
         fn = get_jaxified_graph(inputs=[x], outputs=[y])
-    assert not record
     assert np.isclose(fn(0), 1)
 
 
@@ -120,7 +130,9 @@ def test_get_log_likelihood():
         sigma = pm.HalfNormal("sigma")
         b = pm.Normal("b", a, sigma=sigma, observed=obs_at)
 
-        trace = pm.sample(tune=10, draws=10, chains=2, random_seed=1322)
+        with warnings.catch_warnings():
+            warnings.filterwarnings("ignore", ".*number of samples.*", UserWarning)
+            trace = pm.sample(tune=10, draws=10, chains=2, random_seed=1322)
 
     b_true = trace.log_likelihood.b.values
     a = np.array(trace.posterior.a)
@@ -153,11 +165,14 @@ def test_get_jaxified_logp():
     assert not np.isinf(jax_fn((np.array(5000.0), np.array(5000.0))))
 
 
-@pytest.fixture
-def model_test_idata_kwargs(scope="module"):
-    with pm.Model(coords={"x_coord": ["a", "b"], "x_coord2": [1, 2]}) as m:
+@pytest.fixture(scope="module")
+def model_test_idata_kwargs() -> pm.Model:
+    with pm.Model(
+        coords={"x_coord": ["a", "b"], "x_coord2": [1, 2], "z_coord": ["apple", "banana", "orange"]}
+    ) as m:
         x = pm.Normal("x", shape=(2,), dims=["x_coord"])
-        y = pm.Normal("y", x, observed=[0, 0])
+        _ = pm.Normal("y", x, observed=[0, 0])
+        _ = pm.Normal("z", 0, 1, dims="z_coord")
         pm.ConstantData("constantdata", [1, 2, 3])
         pm.MutableData("mutabledata", 2)
     return m
@@ -184,7 +199,13 @@ def model_test_idata_kwargs(scope="module"):
     ],
 )
 @pytest.mark.parametrize("postprocessing_backend", [None, "cpu"])
-def test_idata_kwargs(model_test_idata_kwargs, sampler, idata_kwargs, postprocessing_backend):
+def test_idata_kwargs(
+    model_test_idata_kwargs: pm.Model,
+    sampler: Callable[..., az.InferenceData],
+    idata_kwargs: Dict[str, Any],
+    postprocessing_backend: Optional[str],
+):
+    idata: Optional[az.InferenceData] = None
     with model_test_idata_kwargs:
         idata = sampler(
             tune=50,
@@ -193,19 +214,31 @@ def test_idata_kwargs(model_test_idata_kwargs, sampler, idata_kwargs, postproces
             idata_kwargs=idata_kwargs,
             postprocessing_backend=postprocessing_backend,
         )
-    assert "constantdata" in idata.constant_data
-    assert "mutabledata" in idata.constant_data
+    assert idata is not None
+    const_data = idata.get("constant_data")
+    assert const_data is not None
+    assert "constantdata" in const_data
+    assert "mutabledata" in const_data
 
     if idata_kwargs.get("log_likelihood", True):
         assert "log_likelihood" in idata
     else:
         assert "log_likelihood" not in idata
 
+    posterior = idata.get("posterior")
+    assert posterior is not None
     x_dim_expected = idata_kwargs.get("dims", model_test_idata_kwargs.RV_dims)["x"][0]
-    assert idata.posterior.x.dims[-1] == x_dim_expected
+    assert x_dim_expected is not None
+    assert posterior["x"].dims[-1] == x_dim_expected
 
     x_coords_expected = idata_kwargs.get("coords", model_test_idata_kwargs.coords)[x_dim_expected]
-    assert list(x_coords_expected) == list(idata.posterior.x.coords[x_dim_expected].values)
+    assert x_coords_expected is not None
+    assert list(x_coords_expected) == list(posterior["x"].coords[x_dim_expected].values)
+
+    assert posterior["z"].dims[2] == "z_coord"
+    assert np.all(
+        posterior["z"].coords["z_coord"].values == np.array(["apple", "banana", "orange"])
+    )
 
 
 def test_get_batched_jittered_initial_points():
@@ -270,3 +303,57 @@ def test_seeding(chains, random_seed, sampler):
     if chains > 1:
         assert np.all(result1.posterior["x"].sel(chain=0) != result1.posterior["x"].sel(chain=1))
         assert np.all(result2.posterior["x"].sel(chain=0) != result2.posterior["x"].sel(chain=1))
+
+
+@pytest.mark.parametrize(
+    "nuts_kwargs",
+    [
+        {"adapt_step_size": False},
+        {"adapt_mass_matrix": True},
+        {"dense_mass": True},
+        {"adapt_step_size": False, "adapt_mass_matrix": True, "dense_mass": True},
+        {"fake-key": "fake-value"},
+    ],
+)
+def test_update_numpyro_nuts_kwargs(nuts_kwargs: Dict[str, Any]):
+    original_kwargs = nuts_kwargs.copy()
+    new_kwargs = _update_numpyro_nuts_kwargs(nuts_kwargs)
+
+    # Maintains original key-value pairs.
+    for k, v in original_kwargs.items():
+        assert new_kwargs[k] == v
+
+    for k, v in _numpyro_nuts_defaults().items():
+        if k not in original_kwargs:
+            assert new_kwargs[k] == v
+
+
+@mock.patch("numpyro.infer.MCMC")
+def test_numpyro_nuts_kwargs_are_used(mocked: mock.MagicMock):
+    mocked.side_effect = MCMC
+
+    step_size = 0.13
+    dense_mass = True
+    adapt_step_size = False
+    target_accept = 0.78
+
+    with pm.Model():
+        pm.Normal("a")
+        sample_numpyro_nuts(
+            10,
+            tune=10,
+            chains=1,
+            target_accept=target_accept,
+            nuts_kwargs={
+                "step_size": step_size,
+                "dense_mass": dense_mass,
+                "adapt_step_size": adapt_step_size,
+            },
+        )
+    mocked.assert_called_once()
+    nuts_sampler = mocked.call_args.args[0]
+    assert nuts_sampler._step_size == step_size
+    assert nuts_sampler._dense_mass == dense_mass
+    assert nuts_sampler._adapt_step_size == adapt_step_size
+    assert nuts_sampler._adapt_mass_matrix
+    assert nuts_sampler._target_accept_prob == target_accept
