@@ -12,6 +12,8 @@
 #   See the License for the specific language governing permissions and
 #   limitations under the License.
 
+import warnings
+
 import aesara
 import numpy as np
 import pytest
@@ -108,10 +110,14 @@ class TestShapesBroadcasting:
         ids=str,
     )
     def test_type_check_raises(self, bad_input):
-        with pytest.raises(TypeError):
-            shapes_broadcasting(bad_input, tuple(), raise_exception=True)
-        with pytest.raises(TypeError):
-            shapes_broadcasting(bad_input, tuple(), raise_exception=False)
+        with warnings.catch_warnings():
+            warnings.filterwarnings(
+                "ignore", ".*ragged nested sequences.*", np.VisibleDeprecationWarning
+            )
+            with pytest.raises(TypeError):
+                shapes_broadcasting(bad_input, tuple(), raise_exception=True)
+            with pytest.raises(TypeError):
+                shapes_broadcasting(bad_input, tuple(), raise_exception=False)
 
     def test_type_check_success(self):
         inputs = [3, 3.0, tuple(), [3], (3,), np.array(3), np.array([3])]
@@ -202,16 +208,14 @@ class TestSamplesBroadcasting:
 
 
 class TestShapeDimsSize:
-    @pytest.mark.parametrize("param_shape", [(), (3,)])
+    @pytest.mark.parametrize("param_shape", [(), (2,)])
     @pytest.mark.parametrize("batch_shape", [(), (3,)])
     @pytest.mark.parametrize(
         "parametrization",
         [
             "implicit",
             "shape",
-            "shape...",
             "dims",
-            "dims...",
             "size",
         ],
     )
@@ -238,7 +242,8 @@ class TestShapeDimsSize:
         with pm.Model(coords=coords) as pmodel:
             mu = aesara.shared(np.random.normal(size=param_shape))
 
-            with pytest.warns(None):
+            with warnings.catch_warnings():
+                warnings.simplefilter("error")
                 if parametrization == "implicit":
                     rv = pm.Normal("rv", mu=mu).shape == param_shape
                 else:
@@ -246,68 +251,59 @@ class TestShapeDimsSize:
                     if parametrization == "shape":
                         rv = pm.Normal("rv", mu=mu, shape=batch_shape + param_shape)
                         assert rv.eval().shape == expected_shape
-                    elif parametrization == "shape...":
-                        rv = pm.Normal("rv", mu=mu, shape=(*batch_shape, ...))
-                        assert rv.eval().shape == batch_shape + param_shape
                     elif parametrization == "dims":
                         rv = pm.Normal("rv", mu=mu, dims=batch_dims + param_dims)
                         assert rv.eval().shape == expected_shape
-                    elif parametrization == "dims...":
-                        rv = pm.Normal("rv", mu=mu, dims=(*batch_dims, ...))
-                        n_size = len(batch_shape)
-                        n_implied = len(param_shape)
-                        ndim = n_size + n_implied
-                        assert len(pmodel.RV_dims["rv"]) == ndim, pmodel.RV_dims
-                        assert len(pmodel.RV_dims["rv"][:n_size]) == len(batch_dims)
-                        assert len(pmodel.RV_dims["rv"][n_size:]) == len(param_dims)
-                        if n_implied > 0:
-                            assert pmodel.RV_dims["rv"][-1] is None
                     elif parametrization == "size":
                         rv = pm.Normal("rv", mu=mu, size=batch_shape + param_shape)
                         assert rv.eval().shape == expected_shape
                     else:
                         raise NotImplementedError("Invalid test case parametrization.")
 
-    @pytest.mark.parametrize("ellipsis_in", ["none", "shape", "dims", "both"])
-    def test_simultaneous_shape_and_dims(self, ellipsis_in):
+    def test_broadcast_by_dims(self):
+        with pm.Model(coords={"broadcast_dim": range(3)}) as m:
+            x = pm.Normal("x", mu=np.zeros((1,)), dims=("broadcast_dim",))
+            assert x.eval().shape == (3,)
+
+    def test_broadcast_by_observed(self):
+        with pm.Model() as m:
+            x = pm.Normal("x", mu=np.zeros((1,)), observed=np.zeros((3,)))
+            assert x.eval().shape == (3,)
+
+    def test_simultaneous_shape_and_dims(self):
         with pm.Model() as pmodel:
             x = pm.ConstantData("x", [1, 2, 3], dims="ddata")
 
-            if ellipsis_in == "none":
-                # The shape and dims tuples correspond to each other.
-                # Note: No checks are performed that implied shape (x), shape and dims actually match.
-                y = pm.Normal("y", mu=x, shape=(2, 3), dims=("dshape", "ddata"))
-                assert pmodel.RV_dims["y"] == ("dshape", "ddata")
-            elif ellipsis_in == "shape":
-                y = pm.Normal("y", mu=x, shape=(2, ...), dims=("dshape", "ddata"))
-                assert pmodel.RV_dims["y"] == ("dshape", "ddata")
-            elif ellipsis_in == "dims":
-                y = pm.Normal("y", mu=x, shape=(2, 3), dims=("dshape", ...))
-                assert pmodel.RV_dims["y"] == ("dshape", None)
-            elif ellipsis_in == "both":
-                y = pm.Normal("y", mu=x, shape=(2, ...), dims=("dshape", ...))
-                assert pmodel.RV_dims["y"] == ("dshape", None)
+            # The shape and dims tuples correspond to each other.
+            # Note: No checks are performed that implied shape (x), shape and dims actually match.
+            y = pm.Normal("y", mu=x, shape=(2, 3), dims=("dshape", "ddata"))
+            assert pmodel.RV_dims["y"] == ("dshape", "ddata")
 
             assert "dshape" in pmodel.dim_lengths
             assert y.eval().shape == (2, 3)
 
-    @pytest.mark.parametrize("with_dims_ellipsis", [False, True])
-    def test_simultaneous_size_and_dims(self, with_dims_ellipsis):
+    def test_simultaneous_size_and_dims(self):
         with pm.Model() as pmodel:
             x = pm.ConstantData("x", [1, 2, 3], dims="ddata")
             assert "ddata" in pmodel.dim_lengths
 
             # Size does not include support dims, so this test must use a dist with support dims.
             kwargs = dict(name="y", size=(2, 3), mu=at.ones((3, 4)), cov=at.eye(4))
-            if with_dims_ellipsis:
-                y = pm.MvNormal(**kwargs, dims=("dsize", ...))
-                assert pmodel.RV_dims["y"] == ("dsize", None, None)
-            else:
-                y = pm.MvNormal(**kwargs, dims=("dsize", "ddata", "dsupport"))
-                assert pmodel.RV_dims["y"] == ("dsize", "ddata", "dsupport")
+            y = pm.MvNormal(**kwargs, dims=("dsize", "ddata", "dsupport"))
+            assert pmodel.RV_dims["y"] == ("dsize", "ddata", "dsupport")
 
             assert "dsize" in pmodel.dim_lengths
             assert y.eval().shape == (2, 3, 4)
+
+    def test_simultaneous_dims_and_observed(self):
+        with pm.Model() as pmodel:
+            x = pm.ConstantData("x", [1, 2, 3], dims="ddata")
+            assert "ddata" in pmodel.dim_lengths
+
+            # Note: No checks are performed that observed and dims actually match.
+            y = pm.Normal("y", observed=[0, 0, 0], dims="ddata")
+            assert pmodel.RV_dims["y"] == ("ddata",)
+            assert y.eval().shape == (3,)
 
     def test_define_dims_on_the_fly(self):
         with pm.Model() as pmodel:
@@ -328,11 +324,26 @@ class TestShapeDimsSize:
             # The change should propagate all the way through
             assert effect.eval().shape == (4,)
 
+    def test_define_dims_on_the_fly_from_observed(self):
+        with pm.Model() as pmodel:
+            data = aesara.shared(np.zeros((4, 5)))
+            x = pm.Normal("x", observed=data, dims=("patient", "trials"))
+            assert pmodel.dim_lengths["patient"].eval() == 4
+            assert pmodel.dim_lengths["trials"].eval() == 5
+
+            # Use dim to create a new RV
+            x_noisy = pm.Normal("x_noisy", 0, dims=("patient", "trials"))
+            assert x_noisy.eval().shape == (4, 5)
+
+            # Change data patient dims
+            data.set_value(np.zeros((10, 6)))
+            assert x_noisy.eval().shape == (10, 6)
+
     def test_can_resize_data_defined_size(self):
         with pm.Model() as pmodel:
             x = pm.MutableData("x", [[1, 2, 3, 4]], dims=("first", "second"))
             y = pm.Normal("y", mu=0, dims=("first", "second"))
-            z = pm.Normal("z", mu=y, observed=np.ones((1, 4)))
+            z = pm.Normal("z", mu=y, observed=np.ones((1, 4)), size=y.shape)
             assert x.eval().shape == (1, 4)
             assert y.eval().shape == (1, 4)
             assert z.eval().shape == (1, 4)
@@ -379,7 +390,6 @@ class TestShapeDimsSize:
             pm.Normal.dist(mu=mu, dims=("town",))
         assert pm.Normal.dist(mu=mu, shape=(3,)).eval().shape == (3,)
         assert pm.Normal.dist(mu=mu, shape=(5, 3)).eval().shape == (5, 3)
-        assert pm.Normal.dist(mu=mu, shape=(7, ...)).eval().shape == (7, 3)
         assert pm.Normal.dist(mu=mu, size=(3,)).eval().shape == (3,)
         assert pm.Normal.dist(mu=mu, size=(4, 3)).eval().shape == (4, 3)
 
@@ -405,10 +415,6 @@ class TestShapeDimsSize:
         assert rv.ndim == 3
         assert tuple(rv.shape.eval()) == (5, 4, 3)
 
-        rv = pm.MvNormal.dist(mu=np.ones((4, 3, 2)), cov=np.eye(2), shape=(6, 5, ...))
-        assert rv.ndim == 5
-        assert tuple(rv.shape.eval()) == (6, 5, 4, 3, 2)
-
         rv = pm.MvNormal.dist(mu=[1, 2, 3], cov=np.eye(3), size=(5, 4))
         assert tuple(rv.shape.eval()) == (5, 4, 3)
 
@@ -419,22 +425,16 @@ class TestShapeDimsSize:
         assert convert_dims(dims="town") == ("town",)
         with pytest.raises(ValueError, match="must be a tuple, str or list"):
             convert_dims(3)
-        with pytest.raises(ValueError, match="may only appear in the last position"):
-            convert_dims(dims=(..., "town"))
 
     def test_convert_shape(self):
         assert convert_shape(5) == (5,)
         with pytest.raises(ValueError, match="tuple, TensorVariable, int or list"):
             convert_shape(shape="notashape")
-        with pytest.raises(ValueError, match="may only appear in the last position"):
-            convert_shape(shape=(3, ..., 2))
 
     def test_convert_size(self):
         assert convert_size(7) == (7,)
         with pytest.raises(ValueError, match="tuple, TensorVariable, int or list"):
             convert_size(size="notasize")
-        with pytest.raises(ValueError, match="cannot contain"):
-            convert_size(size=(3, ...))
 
     def test_lazy_flavors(self):
         assert pm.Uniform.dist(2, [4, 5], size=[3, 2]).eval().shape == (3, 2)
